fixes issue #1075 

The problem is that:
method `handle_mime_body` tries to check `@buf =~ rx` every time `@buf` is changed,
and `@buf =~ rx` is slow because it tries to seach `rx` from the beginning of `@buf` every time.

I changed `@buf =~ rx` to `@buf.index rx, appropriate_offset` to fix this.