@tompng 2017-06-14
Benchmarks

```
# before
1MB: 0.018405 sec
2MB: 0.045491 sec
4MB: 0.132708 sec
8MB: 0.464206 sec
16MB: 1.57744 sec
32MB: 5.183627 sec
64MB: 26.803011 sec
execution expired (HTTPClient::ReceiveTimeoutError)

# after
1MB: 0.022998 sec
2MB: 0.01944 sec
4MB: 0.043587 sec  # x3 faster
8MB: 0.086238 sec
16MB: 0.131328 sec # x12 faster
32MB: 0.234914 sec
64MB: 0.491324 sec # x54 faster
128MB: 1.159996 sec
```


Benchmark codes

Gemfile
```ruby
source 'https://rubygems.org'
ruby '2.4.0'
gem 'sinatra'
gem 'httpclient'
# gem 'rack', git: 'https://github.com/tompng/rack'
```

server
```sh
bundle install
be ruby -e "require 'sinatra';post('/upload'){params[:file][:tempfile].size.to_s}"
```

benchmark
```ruby
require 'httpclient'
def test size
  file = Tempfile.new
  file.write 'a'*size
  file.rewind
  option = { 'content-type' => "multipart/form-data, boundary=#{rand}" }
  t = Time.now
  res = HTTPClient.post('http://localhost:4567/upload', { file: file }, option).body
  raise if res.to_i != size
  Time.now - t
end
[1,2,4,8,16,32,64,128].each do |mb|
  time = test mb*1000*1000
  puts "#{mb}MB: #{time} sec"
end
```